# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecolibary API",
        "description": " API pour gérer une collection de livres classés par catégories. Le système permet de suivre la disponibilité des ouvrages et de surveiller l'état physique de la collection en suivant précisément les exemplaires dégradés.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/books/{bookId}/borrow": {
            "post": {
                "tags": [
                    "Emprunts"
                ],
                "summary": "Emprunter un livre",
                "description": "Permet à un utilisateur authentifié d'emprunter un exemplaire disponible d'un livre. Un utilisateur ne peut pas emprunter le même livre deux fois simultanément.",
                "operationId": "dabcc43ac92762a80673a9faf16eb74c",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "Identifiant unique du livre à emprunter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Livre emprunté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrowed successfully."
                                        },
                                        "borrow": {
                                            "$ref": "#/components/schemas/Borrow"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Impossible d'effectuer l'emprunt (aucun exemplaire disponible ou déjà emprunté)"
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    },
                    "404": {
                        "description": "Livre introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/borrows/{borrowId}/return": {
            "post": {
                "tags": [
                    "Emprunts"
                ],
                "summary": "Retourner un livre emprunté",
                "description": "Permet à un utilisateur de restituer un livre qu'il a précédemment emprunté. Seul le propriétaire de l'emprunt peut effectuer cette action.",
                "operationId": "5033941f0ec9d514dc292f86e7b49ee4",
                "parameters": [
                    {
                        "name": "borrowId",
                        "in": "path",
                        "description": "Identifiant unique de l'emprunt à clôturer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livre restitué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book returned successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ce livre a déjà été restitué"
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    },
                    "403": {
                        "description": "Action non autorisée - Cet emprunt ne vous appartient pas"
                    },
                    "404": {
                        "description": "Emprunt introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/books/statistics/most/viewed": {
            "get": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Livres les plus consultés",
                "description": "Retourne le top 5 des livres les plus consultés dans la bibliothèque, classés par nombre de vues décroissant.",
                "operationId": "14ca4641c6977e5898702002d00a7013",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "most_viewed_books": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/statistics/most/borrowed": {
            "get": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Livres les plus empruntés",
                "description": "Retourne le top 5 des livres les plus empruntés dans la bibliothèque, classés par nombre d'emprunts décroissant.",
                "operationId": "af82e98bad4b37ccf1d03cd25251d36b",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "most_borrowed_books": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié - Veuillez vous connecter d'abord"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book",
                "description": "Modèle représentant un livre dans la bibliothèque EcoLibrary",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Le Petit Prince"
                    },
                    "author": {
                        "type": "string",
                        "example": "Antoine de Saint-Exupéry"
                    },
                    "slug": {
                        "type": "string",
                        "example": "le-petit-prince"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un conte poétique et philosophique sous l'apparence d'un livre pour enfants.",
                        "nullable": true
                    },
                    "total_copies": {
                        "type": "integer",
                        "example": 10
                    },
                    "available_copies": {
                        "type": "integer",
                        "example": 7
                    },
                    "degraded_copies": {
                        "type": "integer",
                        "example": 1
                    },
                    "views": {
                        "type": "integer",
                        "example": 152
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    }
                },
                "type": "object"
            },
            "Borrow": {
                "title": "Borrow",
                "description": "Modèle représentant un emprunt de livre",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "book_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "borrowed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-14T10:00:00Z"
                    },
                    "returned_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "en cours",
                            "returned"
                        ],
                        "example": "en cours"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Model for book categories",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Science Fiction"
                    },
                    "slug": {
                        "description": "URL-friendly slug of the category",
                        "type": "string",
                        "example": "science-fiction"
                    },
                    "created_at": {
                        "description": "Timestamp when the category was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the category was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}